modularize init routine for renderer vs embedded use
